+++ flake.nix
@@ -259,11 +259,11 @@
       ref   = "r1.0.3.7";
       flake = false;
     };
-    mockio-cmds-inetutils-src-1-0-0-1 = {
+    mockio-cmds-inetutils-src-1-0-0-2 = {
       type  = "github";
       owner = "sixears";
       repo  = "mockio-cmds-inetutils";
-      ref   = "r1.0.0.1";
+      ref   = "r1.0.0.2";
       flake = false;
     };
     mockio-cmds-rsync-src-1-0-0-1 = {
@@ -528,7 +528,7 @@
             , log-plus-src-0-0-4-4
             , mac-address-src-0-0-1-0
             , minfo-src-1-0-3-7
-            , mockio-cmds-inetutils-src-1-0-0-1
+            , mockio-cmds-inetutils-src-1-0-0-2
             , mockio-cmds-rsync-src-1-0-0-1
             , mockio-cmds-util-linux-src-1-0-1-3
             , mockio-src-0-0-6-0
@@ -1662,17 +1662,15 @@
 
           # -- mockio-cmds-inetutils ---
 
-          mockio-cmds-inetutils         = mockio-cmds-inetutils-1-0;
-          mockio-cmds-inetutils-1-0     = mockio-cmds-inetutils-1-0-0-1;
-          mockio-cmds-inetutils-1-0-0-1 =
-            callPkg "mockio-cmds-inetutils" "1.0.0.1"
-                    mockio-cmds-inetutils-src-1-0-0-1 {
+          mockio-cmds-inetutils-1-0-0-2 =
+            callPkg "mockio-cmds-inetutils" "1.0.0.2"
+                    mockio-cmds-inetutils-src-1-0-0-2 {
             description = "MockIO wrappers for inetutils cmds";
             libDepends = h: with h; [
               data-textual lens logging-effect mtl safe text
 
               base1t domainnames fpath log-plus mockio mockio-log mockio-plus
-              monadio-plus l0.more-unicode stdmain
+              monadio-plus stdmain
             ];
 
             postConfigure = ''
@@ -1683,6 +1681,9 @@
                 done
             '';
           };
+
+          mockio-cmds-inetutils-1-0     = mockio-cmds-inetutils-1-0-0-2;
+          mockio-cmds-inetutils         = mockio-cmds-inetutils-1-0;
 
           # -- mockio-cmds-rsync -------